Yak shaving: redundant casts

+++ jmh-core/src/test/java/org/openjdk/jmh/util/TestListStatistics.java
@@ -76,7 +76,7 @@ public class TestListStatistics {
      */
     @Test
     public strictfp void testGetN() {
-        assertEquals((long) VALUES.length, (long) instance.getN());
+        assertEquals((long) VALUES.length, instance.getN());
     }
 
     /**
