+++ jmh-core/src/test/java/org/openjdk/jmh/util/TestMultisetStatistics.java
@@ -77,7 +77,7 @@ public class TestMultisetStatistics {
      */
     @Test
     public strictfp void testGetN() {
-        assertEquals((long) VALUES.length, (long) instance.getN());
+        assertEquals((long) VALUES.length, instance.getN());
     }
 
     /**
